@@ -19,7 +19,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ^3.5.3
+  sdk: ^3.6.1
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -32,22 +32,32 @@ dependencies:
     sdk: flutter
   flutter_localizations:
     sdk: flutter
-
+  flutter_spinkit: ^5.2.1
+  get: ^4.6.6 
+  flutter_dotenv: ^5.2.1
+  http: ^0.13.6
+  shared_preferences: ^2.5.2
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.8
-  flutter_dotenv: ^5.2.1
-  get: ^4.6.6
   salomon_bottom_bar: ^3.3.2
-  flutter_riverpod: ^2.6.1
-  animate_do: ^3.3.4
-  http: ^1.2.2
-  quickalert: ^1.1.0
-  shared_preferences: ^2.3.2
-  flutter_spinkit: ^5.2.1
   socket_io_client: ^2.0.0
   freezed_annotation: ^2.0.0
-  freezed: ^2.0.0
+
+  cupertino_icons: ^1.0.8
+  flutter_riverpod: ^2.6.1
+  animate_do: ^3.3.4
+  quickalert: ^1.1.0
+  geolocator: ^13.0.4
+  geocoding: ^3.0.0
+
+
+flutter_launcher_icons:
+  android: true
+  ios: true
+  image_path: "assets/bcg_store.png" 
+  adaptive_icon_background: "#3F3F3F" 
+  adaptive_icon_foreground: "assets/bcg_store.png"
+
 dev_dependencies:
   flutter_test:
     sdk: flutter
@@ -57,7 +67,7 @@ dev_dependencies:
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^4.0.0
+  flutter_lints: ^5.0.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
@@ -65,18 +75,20 @@ dev_dependencies:
 # The following section is specific to Flutter packages.
 flutter:
   assets:
+
      - assets/img/
      - assets/language/
      - .env
-     - dev.env
      - test.env
+     - dev.env
   # The following line ensures that the Material Icons font is
   # included with your application, so that you can use the icons in
   # the material Icons class.
   uses-material-design: true
 
   # To add assets to your application, add an assets section, like this:
-
+  # assets:
+  #   - images/a_dot_burr.jpeg
   #   - images/a_dot_ham.jpeg
 
   # An image asset can refer to one or more resolution-specific "variants", see
