@@ -1,8 +1,17 @@
 class AuthEntitie {
   final String? uuid;
-  final String? token;
   final String? correo;
   final String? contrasena;
-  final String? dato;
-  AuthEntitie({this.uuid, this.correo, this.contrasena, this.token, this.dato});
+  final dynamic persona;
+  final dynamic userOrAdmin;
+  final String? token;
+
+  AuthEntitie({
+     this.uuid,
+     this.correo,
+     this.contrasena,
+    this.persona,
+    this.userOrAdmin,
+    this.token,
+  });
 }
