@@ -1,3 +1,4 @@
+import 'package:aleralarma/common/theme/app_theme.dart';
 import 'package:aleralarma/features/auth/presentation/page/login/login_controller.dart';
 import 'package:animate_do/animate_do.dart';
 import 'package:flutter/material.dart';
@@ -63,225 +64,204 @@ class LoginPage2 extends ConsumerWidget {
   Widget build(BuildContext context, WidgetRef ref) {
     final authState = ref.watch(authProvider);
 
-    return Scaffold(
-      body: Stack(
-        children: [
-          FadeInDown(
-            duration: const Duration(milliseconds: 1200),
-            child: ClipPath(
-              clipper: TopShapeClipper(),
-              child: Container(
-                height: MediaQuery.of(context).size.height * 0.2,
-                decoration: const BoxDecoration(
-                  gradient: LinearGradient(
-                    colors: [
-                      Color(0xFF1F2937),
-                      Color(0xFF1F2937),
-                    ],
-                    begin: Alignment.topCenter,
-                    end: Alignment.bottomCenter,
-                  ),
-                ),
-              ),
-            ),
-          ),
-
-          // Curva inferior
-          Positioned(
-            bottom:
-                -MediaQuery.of(context).size.height * 0.1, // Ajuste de posición
-            left: 0,
-            right: 0,
-            child: FadeInUp(
+    return Theme(
+      data: AppTheme.themeData,
+      child: Scaffold(
+        body: Stack(
+          children: [
+            // Curva superior
+            FadeInDown(
               duration: const Duration(milliseconds: 1200),
               child: ClipPath(
-                clipper: BottomShapeClipper(),
+                clipper: TopShapeClipper(),
                 child: Container(
-                  height: MediaQuery.of(context).size.height *
-                      0.2, // Aumentado a 0.4
-                  decoration: const BoxDecoration(
+                  height: MediaQuery.of(context).size.height * 0.2,
+                  decoration: BoxDecoration(
                     gradient: LinearGradient(
                       colors: [
-                        Color(0xFF1F2937),
-                        Color(0xFF1F2937),
+                        AppTheme.primaryColor,
+                        AppTheme.primaryColor,
                       ],
-                      begin: Alignment.bottomCenter,
-                      end: Alignment.topCenter,
+                      begin: Alignment.topCenter,
+                      end: Alignment.bottomCenter,
                     ),
                   ),
                 ),
               ),
             ),
-          ),
 
-          // Contenido principal
-          SafeArea(
-            child: Column(
-              children: [
-                // Header animado
-                FadeInDown(
-                  duration: const Duration(milliseconds: 1000),
-                  child: Padding(
-                    padding: const EdgeInsets.symmetric(horizontal: 8.0),
-                    child: Row(
-                      children: [
-                        IconButton(
-                          icon:
-                              const Icon(Icons.arrow_back, color: Colors.white),
-                          onPressed: () => Navigator.pop(context),
-                        ),
-                        const Expanded(
-                          child: Center(
-                            child: Text(
-                              'LOGIN',
-                              style: TextStyle(
-                                color: Color(0xFFB8860B),
-                                fontSize: 20,
-                                fontWeight: FontWeight.bold,
+            // Curva inferior
+            Positioned(
+              bottom: -MediaQuery.of(context).size.height * 0.1,
+              left: 0,
+              right: 0,
+              child: FadeInUp(
+                duration: const Duration(milliseconds: 1200),
+                child: ClipPath(
+                  clipper: BottomShapeClipper(),
+                  child: Container(
+                    height: MediaQuery.of(context).size.height * 0.2,
+                    decoration: BoxDecoration(
+                      gradient: LinearGradient(
+                        colors: [
+                          AppTheme.primaryColor,
+                          AppTheme.primaryColor,
+                        ],
+                        begin: Alignment.bottomCenter,
+                        end: Alignment.topCenter,
+                      ),
+                    ),
+                  ),
+                ),
+              ),
+            ),
+
+            // Contenido principal
+            SafeArea(
+              child: Column(
+                children: [
+                  // Header animado
+                  FadeInDown(
+                    duration: const Duration(milliseconds: 1000),
+                    child: Padding(
+                      padding: const EdgeInsets.symmetric(horizontal: 8.0),
+                      child: Row(
+                        children: [
+                         
+                          Expanded(
+                            child: Center(
+                              child: Text(
+                                'LOGIN',
+                                style: TextStyle(
+                                  color: AppTheme.accentColor,
+                                  fontSize: 20,
+                                  fontWeight: FontWeight.bold,
+                                ),
                               ),
                             ),
                           ),
-                        ),
-                        const SizedBox(width: 48),
-                      ],
+                          const SizedBox(width: 48),
+                        ],
+                      ),
                     ),
                   ),
-                ),
 
-                // Contenido del formulario centrado
-                Expanded(
-                  child: Center(
-                    child: SingleChildScrollView(
-                      child: FadeIn(
-                        duration: const Duration(milliseconds: 1500),
-                        child: Container(
-                          margin: const EdgeInsets.symmetric(horizontal: 20),
-                          padding: const EdgeInsets.all(20),
-                          width: double.infinity,
-                          constraints: const BoxConstraints(
-                            maxWidth: 400,
-                          ),
-                          decoration: BoxDecoration(
-                            color: Colors.transparent,
-                            borderRadius: BorderRadius.circular(20),
-                            boxShadow: [],
-                          ),
-                          child: Column(
-                            mainAxisSize: MainAxisSize.min,
-                            children: [
-                              // Logo con bounce
-                              BounceInDown(
-                                duration: const Duration(milliseconds: 1500),
-                                child: Image.asset(
-                                  'assets/img/icono-escudo.png',
-                                  width: 120,
-                                  height: 120,
-                                ),
-                              ),
-
-                              const SizedBox(height: 20),
-
-                              // Texto de bienvenida con fade
-                              FadeInLeft(
-                                duration: const Duration(milliseconds: 1000),
-                                child: const Text(
-                                  'Bienvenido',
-                                  style: TextStyle(
-                                    color: Colors.blue,
-                                    fontSize: 24,
-                                    fontWeight: FontWeight.bold,
+                  // Contenido del formulario centrado
+                  Expanded(
+                    child: Center(
+                      child: SingleChildScrollView(
+                        child: FadeIn(
+                          duration: const Duration(milliseconds: 1500),
+                          child: Container(
+                            margin: const EdgeInsets.symmetric(horizontal: 20),
+                            padding: const EdgeInsets.all(20),
+                            width: double.infinity,
+                            constraints: const BoxConstraints(
+                              maxWidth: 400,
+                            ),
+                            decoration: BoxDecoration(
+                              color: Colors.transparent,
+                              borderRadius: BorderRadius.circular(20),
+                              boxShadow: [],
+                            ),
+                            child: Column(
+                              mainAxisSize: MainAxisSize.min,
+                              children: [
+                                // Logo con bounce
+                                BounceInDown(
+                                  duration: const Duration(milliseconds: 1500),
+                                  child: Image.asset(
+                                    'assets/img/icono-escudo.png',
+                                    width: 120,
+                                    height: 120,
                                   ),
                                 ),
-                              ),
 
-                              const SizedBox(height: 20),
+                                const SizedBox(height: 20),
 
-                              // Campo de usuario con fade
-                              FadeInLeft(
-                                delay: const Duration(milliseconds: 200),
-                                duration: const Duration(milliseconds: 1000),
-                                child: TextField(
-                                  controller: authState.emailController,
-                                  decoration: InputDecoration(
+                                // Texto de bienvenida con fade
+                                FadeInLeft(
+                                  duration: const Duration(milliseconds: 1000),
+                                  child: Text(
+                                    'Bienvenido',
+                                    style: TextStyle(
+                                      color: AppTheme.communicationColor,
+                                      fontSize: 24,
+                                      fontWeight: FontWeight.bold,
+                                    ),
+                                  ),
+                                ),
+
+                                const SizedBox(height: 20),
+
+                                // Campo de usuario con fade
+                                FadeInLeft(
+                                  delay: const Duration(milliseconds: 200),
+                                  duration: const Duration(milliseconds: 1000),
+                                  child: AppTheme.createTextField(
+                                    controller: authState.emailController,
                                     labelText: 'Usuario',
-                                    prefixIcon: const Icon(Icons.person),
-                                    border: OutlineInputBorder(
-                                      borderRadius: BorderRadius.circular(10),
-                                    ),
+                                    prefixIcon: Icons.person,
+                                    textInputAction: TextInputAction.next,
                                   ),
                                 ),
-                              ),
 
-                              const SizedBox(height: 15),
+                                const SizedBox(height: 15),
 
-                              // Campo de contraseña con fade
-                              FadeInLeft(
-                                delay: const Duration(milliseconds: 400),
-                                duration: const Duration(milliseconds: 1000),
-                                child: TextField(
-                                  controller: authState.passwordController,
-                                  obscureText: true,
-                                  decoration: InputDecoration(
+                                // Campo de contraseña con fade
+                                FadeInLeft(
+                                  delay: const Duration(milliseconds: 400),
+                                  duration: const Duration(milliseconds: 1000),
+                                  child: AppTheme.createTextField(
+                                    controller: authState.passwordController,
                                     labelText: 'Contraseña',
-                                    prefixIcon: const Icon(Icons.lock),
-                                    border: OutlineInputBorder(
-                                      borderRadius: BorderRadius.circular(10),
-                                    ),
-                                  ),
-                                ),
-                              ),
-
-                              const SizedBox(height: 30),
-
-                              // Botón con bounce
-                              BounceInUp(
-                                delay: const Duration(milliseconds: 600),
-                                duration: const Duration(milliseconds: 1000),
-                                child: SizedBox(
-                                  width: double.infinity,
-                                  height: 50,
-                                  child: ElevatedButton(
-                                    onPressed: () {
-                                      ref
-                                          .read(authProvider.notifier)
-                                          .login(context);
+                                    prefixIcon: Icons.lock,
+                                    obscureText: true,
+                                    textInputAction: TextInputAction.done,
+                                    onSubmitted: (_) {
+                                      ref.read(authProvider.notifier).login(context);
                                     },
-                                    style: ElevatedButton.styleFrom(
-                                      backgroundColor: const Color(0xFF1F2937),
-                                      shape: RoundedRectangleBorder(
-                                        borderRadius: BorderRadius.circular(10),
-                                      ),
-                                    ),
-                                    child: const Text(
-                                      'Login',
-                                      style: TextStyle(
-                                        color: Color(0xFFB8860B),
-                                        fontSize: 16,
-                                      ),
-                                    ),
                                   ),
                                 ),
-                              ),
-                            ],
+
+                                const SizedBox(height: 30),
+
+                                // Botón con bounce
+                                BounceInUp(
+                                  delay: const Duration(milliseconds: 600),
+                                  duration: const Duration(milliseconds: 1000),
+                                  child: AppTheme.createPrimaryButton(
+                                    text: 'Login',
+                                    onPressed: () {
+                                      ref.read(authProvider.notifier).login(context);
+                                    },
+                                    height: 50,
+                                  ),
+                                ),
+                              ],
+                            ),
                           ),
                         ),
                       ),
                     ),
                   ),
-                ),
-              ],
-            ),
-          ),if (authState.isLoading)
-            Container(
-              color: Colors.black.withOpacity(0.5),
-              child: const Center(
-                child: SpinKitFadingCube(
-                  color: Colors.black,
-                  size: 50.0,
-                ),
+                ],
               ),
             ),
-        ],
+            
+            // Overlay de carga
+            if (authState.isLoading)
+              Container(
+                color: Colors.black.withOpacity(0.5),
+                child: Center(
+                  child: SpinKitFadingCube(
+                    color: AppTheme.accentColor,
+                    size: 50.0,
+                  ),
+                ),
+              ),
+          ],
+        ),
       ),
     );
   }
